minor bugfixes with gcp

+++ portfolio/src/main/java/com/google/sps/servlets/GetCommentsServlet.java
@@ -85,8 +85,9 @@ public class GetCommentsServlet extends HttpServlet {
       }
       long timestamp = (long) entity.getProperty("timestamp");
       double score = (double) entity.getProperty("score");
-      int likes = (int) entity.getProperty("likes");
-      int dislikes = (int) entity.getProperty("dislikes");
+
+      int likes = (int) (long) entity.getProperty("likes");
+      int dislikes = (int) (long) entity.getProperty("dislikes");
 
       Comment comment = new Comment(id, message, timestamp, score, likes, dislikes);
       comments.add(comment);
